requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==8.0.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
